pyproject.toml
[project]
name = "jaximus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "array-record>=0.7.2",
    "datasets>=4.0.0",
    "grain>=0.2.11",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "optax>=0.2.5",
    "orbax-checkpoint>=0.11.21",
    "polars>=1.33.1",
    "sws-config>=0.6.2",
    "transformers>=4.55.0",
    "wandb>=0.21.1",
    "xprof>=2.20.6",
]

[project.optional-dependencies]
cpu = ["jax>=0.7.0"]
gpu = ["jax[cuda12]>=0.7.0"]
tpu = ["jax[tpu]>=0.7.0"]
dev = ["ruff>=0.8.0"]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E731", "E741"]
